backend-react/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Elasticsearch
elasticsearch>=8.0.0

# API and Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.27.0
